Moved files around with better names. Added more docstrings.
