+++ software/bootloader.py
@@ -1,53 +1,34 @@
-from europi import *
+"""
+Default EuroPi main entry point which loads a menu of scripts available.
+
+Copy this file to main.py on your EuroPi pico to install the menu bootloader.
+
+When a script is selected, subsequent boots will automatically launch that 
+script. To return to the bootloader menu, press both buttons simultaneously.
+
+"""
 from machine import reset
-import utime as time
+from europi import b1, b2
 import os
 
-# In order for a script to be included in the menu, it must be imported here
-from contrib.coin_toss import CoinToss
-from contrib.consequencer import Consequencer
-from contrib.diagnostic import Diagnostic
-from contrib.harmonic_lfos import HarmonicLFOs
-from contrib.radio_scanner import RadioScanner
-from contrib.polyrhythmic_sequencer import PolyrhythmSeq
+def reset_menu():
+    os.remove("menu_state.py")
+    reset()
+
+b1.handler_both(b2, reset_menu)
+b2.handler_both(b1, reset_menu)
 
 
-# Load all imported scripts that subclass EuroPiScript.
-SCRIPTS = [
-    CoinToss,
-    Consequencer,
-    Diagnostic,
-    HarmonicLFOs,
-    RadioScanner,
-    PolyrhythmSeq,
-]
+# If a previously selected script was saved, load it and execute it.
+try:
+    from menu_state import MODULE_PATH, CLASS_NAME
+    module = __import__(MODULE_PATH, globals(), locals(), [CLASS_NAME])
+    cls = getattr(module, CLASS_NAME)
+    script = cls()
+    script.main()
+except ImportError:
+    pass  # No previous script available, proceed to menu.
 
-
-def display_choice(cls):
-    oled.centre_text(cls.__qualname__)
-
-
-def save_choice(cls):
-    with open(f'previously_selected.py', 'w') as file:
-        file.write(f"MODULE_PATH=\"{cls.__module__}\"\nCLASS_NAME=\"{cls.__name__}\"\n")
-
-def launch_menu():
-
-    # Validate all scripts implement EuroPiScript or fail loudly.
-    # for script in SCRIPTS:
-    #     if not issubclass(script, EuroPiScript):
-    #         raise NotImplementedError(f"Script does not implement EuroPiScript: {script}")
-
-    # Bootloader script selection.
-    while True:
-        cls = k1.choice(SCRIPTS)
-        display_choice(cls)
-
-        # If button 1 is pressed, execute the currently selected script.
-        if b1.value() == 1 or b2.value() == 1:
-            # Save script path to previously_selected.py.
-            save_choice(cls)
-            # Reset to clear memory and launch saved script.
-            reset()
-
-        time.sleep_ms(100)
+# Otherwise, launch the menu
+from bootloader import launch_menu
+launch_menu()
